requirements.txt
# Основные зависимости для бота ремонта компьютеров

# Telegram Bot Framework
aiogram==3.4.1

# База данных
aiosqlite==0.19.0

# ИИ интеграция
google-generativeai==0.8.3

# Дополнительные утилиты
python-dateutil==2.8.2
pytz==2023.3

# Для улучшенного логирования (опционально)
colorlog==6.7.0

# Для работы с конфигурацией (опционально)
python-dotenv==1.0.0

# Для валидации данных (опционально)
pydantic==2.5.3

# Для тестирования (опционально)
pytest==7.4.3
pytest-asyncio==0.21.1

# Для форматирования кода (опционально для разработки)
black==23.11.0
isort==5.12.0
flake8==6.1.0